requirements.txt
# Requirements for Apple Detection Project
# Compatible with Python 3.8+ (tested with Python 3.13.7)
# For CPU-only systems, install PyTorch separately:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Data Processing
pyyaml>=6.0
tqdm>=4.64.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Optional: Data Augmentation
albumentations>=1.3.0

# Optional: Experiment Tracking
# wandb>=0.13.0
# tensorboard>=2.10.0

# Development Tools (Optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0
# black>=22.0.0
# pylint>=2.15.0

